implement processing and creating vector from rasters

+++ processing/utils.py
@@ -7,7 +7,32 @@ from osgeo import gdal
 from osgeo import osr
 from osgeo import ogr
 import os
-import config
+from config.settings import Config
+
+
+def parse_file_name(filename: str, parameter) -> str:
+    d = {
+        "model": 0,
+        "date": 1,
+        "hour": 2,
+        "indicator": 3,
+        "resolution": 4,
+    }
+    idx = d[parameter]
+
+    return filename.split(".")[idx]
+
+
+def extract_rasters(inArchive, outPath, indicatorsForExtract):
+
+    from zipfile import ZipFile
+
+    with ZipFile(inArchive, 'r') as zipObject:
+        listOfFileNames = zipObject.namelist()
+        for fileName in listOfFileNames:
+            indicator = parse_file_name(fileName, "indicator")
+            if indicator in indicatorsForExtract:
+                zipObject.extract(fileName, outPath)
 
 
 def raster_2_binary(rasterPath: str, function) -> np.array:
@@ -58,9 +83,9 @@ def create_template_raster(outPath):
     outSpatialRef = osr.SpatialReference()
     outSpatialRef.ImportFromEPSG(4326)
     driver = gdal.GetDriverByName("GTiff")
-    dstDs = driver.Create(outPath, config.RASTER_X_SIZE, config.RASTER_Y_SIZE, 1, gdal.GDT_Byte)
+    dstDs = driver.Create(outPath, Config.RASTER_X_SIZE, Config.RASTER_Y_SIZE, 1, gdal.GDT_Byte)
     dstDs.SetProjection(outSpatialRef.ExportToWkt())
-    dstDs.SetGeoTransform(config.RASTER_GEO_TRANSFORM)
+    dstDs.SetGeoTransform(Config.RASTER_GEO_TRANSFORM)
 
 
 def check_new_zips(url, dwnFld, startDate=datetime(2021, 6, 19)):
@@ -100,5 +125,17 @@ def download_file_util(baseUrl, zipName, model):
         os.mkdir(save_path)
 
     save_path_zip = os.path.join(save_path, f'{zipName}')
+    #TODO check this
+    if os.path.exists(save_path_zip):
+        return save_path_zip
+
     with urllib.request.urlopen(url) as response, open(save_path_zip, 'wb') as out_file:
         shutil.copyfileobj(response, out_file)
+
+    return save_path_zip
+
+
+def merge_file_name(model, date, hour, indicator) -> str:
+    return f'{model}.{date}.{hour}.{indicator}.tif'
+
+
